metadater/requirements.txt
# ChatGPT Metadater
httpx>=0.25.0
python-dotenv>=1.0.0

